[ENH, MRG] Add decimate epochs mixin
